# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Resource Management and Booking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: department_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: is_active
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /admin/users/{id}/role:
    patch:
      operationId: adminUpdateUserRole
      tags: [admin]
      summary: Update user role, department, and active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Soft delete user (deactivate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ADMIN - Departments
  /admin/departments:
    get:
      operationId: adminListDepartments
      tags: [admin]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepartmentDetail"
    post:
      operationId: adminCreateDepartment
      tags: [admin]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentDetail"
  /admin/departments/{id}:
    patch:
      operationId: adminUpdateDepartment
      tags: [admin]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentDetail"

  # ADMIN - Blackout
  /admin/blackout:
    get:
      operationId: adminListBlackout
      tags: [admin]
      summary: List all blackout periods
      responses:
        "200":
          description: List of blackout periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlackoutPeriod"
    post:
      operationId: adminCreateBlackout
      tags: [admin]
      summary: Create blackout period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlackoutBody"
      responses:
        "201":
          description: Created blackout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlackoutPeriod"
  /admin/blackout/{id}:
    delete:
      operationId: adminDeleteBlackout
      tags: [admin]
      summary: Delete blackout period
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/audit-log:
    get:
      operationId: adminGetAuditLog
      tags: [admin]
      summary: Get audit log
      parameters:
        - name: table_name
          in: query
          schema:
            type: string
        - name: operation
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Paginated audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAuditLog"
  /admin/bookings:
    get:
      operationId: adminListBookings
      tags: [admin]
      summary: List all bookings system-wide
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: resource_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: user_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: from_date
          in: query
          schema:
            type: string
        - name: to_date
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Paginated bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBookings"
  /admin/bookings/{id}/cancel:
    patch:
      operationId: adminCancelBooking
      tags: [admin]
      summary: Cancel any booking (admin override)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        user_id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        is_active:
          type: boolean
        priority_level:
          type: integer
        department_id:
          type: ["integer", "null"]
        department_name:
          type: ["string", "null"]
        created_at:
          type: string
      required: [user_id, first_name, last_name, email, role, is_active, priority_level]
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
        is_active:
          type: boolean
        department_id:
          type: ["integer", "null"]
      required: [role]
...
    DepartmentDetail:
      type: object
      properties:
        department_id:
          type: integer
        dept_name:
          type: string
        building:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hod_id:
          type: ["integer", "null"]
        hod_name:
          type: ["string", "null"]
        user_count:
          type: integer
      required: [department_id, dept_name, user_count]
    CreateDepartmentBody:
      type: object
      properties:
        dept_name:
          type: string
        building:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hod_id:
          type: ["integer", "null"]
      required: [dept_name]
    BlackoutPeriod:
      type: object
      properties:
        blackout_id:
          type: integer
        category_id:
          type: ["integer", "null"]
        category_name:
          type: ["string", "null"]
        start_date:
          type: string
        end_date:
          type: string
        reason:
          type: ["string", "null"]
      required: [blackout_id, start_date, end_date]
    CreateBlackoutBody:
      type: object
      properties:
        category_id:
          type: ["integer", "null"]
        start_date:
          type: string
        end_date:
          type: string
        reason:
          type: ["string", "null"]
      required: [start_date, end_date]
    AuditLogEntry:
      type: object
      properties:
        audit_id:
          type: integer
        table_name:
          type: string
        operation:
          type: string
        old_data:
          type: ["object", "null"]
        new_data:
          type: ["object", "null"]
        changed_by:
          type: ["integer", "null"]
        changed_by_name:
          type: ["string", "null"]
        changed_at:
          type: string
      required: [audit_id, table_name, operation, changed_at]
    PaginatedAuditLog:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    BookingDetail:
      type: object
      properties:
        booking_id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        resource_id:
          type: integer
        resource_name:
          type: string
        category_name:
          type: string
        status_id:
          type: integer
        status_name:
          type: string
        priority_level:
          type: integer
        date:
          type: string
        start_time:
          type: string
        end_time:
          type: string
        purpose:
          type: ["string", "null"]
        created_at:
          type: string
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRecord"
      required: [booking_id, user_id, user_name, resource_id, resource_name, category_name, status_id, status_name, priority_level, date, start_time, end_time, created_at, approvals]
    ApprovalRecord:
      type: object
      properties:
        approval_id:
          type: integer
        step_number:
          type: integer
        decision:
          type: ["string", "null"]
        approver_name:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        approval_time:
          type: ["string", "null"]
      required: [approval_id, step_number]
    PaginatedBookings:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BookingDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
...